updated AF2-gateway.json
Currently not managing to deploy AF2 to prod. I think the main reason for this is that the git repo is written in capitals 'AF2' and as a result the docker image that is being created has the reference quay.io/labdao/AF2@she256...
Putting quay.io/labdao/af2@she256..., i.e. lower case letters in the tool manifest is not consistent and is going to cause problems. Furthermore, I added comma in the json file which in my opinion was missing.